fixed addFood function, renamed the foodData function
